requirements.txt
django
django-crispy-forms==1.11.2
django-filter